small tweak to eliminate u64 to usize casting and updating to 1.58.1 per sec patch

+++ core/lib/src/lib.rs
@@ -62,13 +62,13 @@ pub fn decode_tag_and_data<'a, T: From<&'a [u8]>>(encoded: &[u8], buf: &'a mut [
     ENCODER.decode_mut(encoded, &mut buf[0..len]).map_err(|_| Error::DecodeError)?;
     let tag = TagBuilder::from_bytes(&buf).build()?;
     let len = tag.len();
-    let data_len = tag.get_data_length() as usize;
+    let data_len = tag.get_data_length();
     let data = T::from(&buf[len..len + data_len]);
     Ok((tag, data))
 }
 
 pub fn encode_tag_and_data<'a>(tag: &mut Tag, data: &impl CryptoData, buf: &'a mut [u8]) -> Result<usize> {
-    tag.set_data_length(data.len() as u64);
+    tag.set_data_length(data.len());
     let tagsize = tag.encode(buf);
     let datasize = data.encode(&mut buf[tagsize..]);
     Ok(tagsize + datasize)
